% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%l0\%}
\alias{\%l0\%}
\alias{\%||\%}
\alias{\%@\%}
\alias{\%nin\%}
\title{Infix length 0}
\usage{
x \%l0\% y

x \%||\% y

x \%@\% name

x \%nin\% table
}
\description{
Infix length 0

Infix length is.null

Infix attr

Infix not-in
}
