% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upfront-package.R
\docType{package}
\name{upfront}
\alias{upfront}
\alias{upfront-package}
\title{Base R Convenience Wrappers That Put Data Up Front}
\description{
A series of functions that help use the new built-in pipe operator
in R without the need for the heft of the tidyverse.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrbrmstr/upfront}
  \item Report bugs at \url{https://github.com/hrbrmstr/upfront/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
