% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upfront-package.R
\docType{package}
\name{upfront}
\alias{upfront}
\alias{upfront-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/upfront}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/upfront/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
